loadbalancer/zk: test cleanups
