+++ loadbalancer/zk/client_test.go
@@ -6,6 +6,8 @@ import (
 	"time"
 
 	stdzk "github.com/samuel/go-zookeeper/zk"
+
+	"github.com/go-kit/kit/log"
 )
 
 func TestNewClient(t *testing.T) {
@@ -18,13 +20,12 @@ func TestNewClient(t *testing.T) {
 
 	c, err := NewClient(
 		[]string{"FailThisInvalidHost!!!"},
-		logger,
+		log.NewNopLogger(),
 	)
-
-	time.Sleep(1 * time.Millisecond)
 	if err == nil {
 		t.Errorf("expected error, got nil")
 	}
+
 	hasFired := false
 	calledEventHandler := make(chan struct{})
 	eventHandler := func(event stdzk.Event) {
@@ -34,9 +35,10 @@ func TestNewClient(t *testing.T) {
 			close(calledEventHandler)
 		}
 	}
+
 	c, err = NewClient(
 		[]string{"localhost"},
-		logger,
+		log.NewNopLogger(),
 		ACL(acl),
 		ConnectTimeout(connectTimeout),
 		SessionTimeout(sessionTimeout),
@@ -47,9 +49,10 @@ func TestNewClient(t *testing.T) {
 		t.Fatal(err)
 	}
 	defer c.Stop()
+
 	clientImpl, ok := c.(*client)
 	if !ok {
-		t.Errorf("retrieved incorrect Client implementation")
+		t.Fatal("retrieved incorrect Client implementation")
 	}
 	if want, have := acl, clientImpl.acl; want[0] != have[0] {
 		t.Errorf("want %+v, have %+v", want, have)
@@ -72,22 +75,22 @@ func TestNewClient(t *testing.T) {
 }
 
 func TestOptions(t *testing.T) {
-	_, err := NewClient([]string{"localhost"}, logger, Credentials("valid", "credentials"))
+	_, err := NewClient([]string{"localhost"}, log.NewNopLogger(), Credentials("valid", "credentials"))
 	if err != nil && err != stdzk.ErrNoServer {
 		t.Errorf("unexpected error: %v", err)
 	}
 
-	_, err = NewClient([]string{"localhost"}, logger, Credentials("nopass", ""))
+	_, err = NewClient([]string{"localhost"}, log.NewNopLogger(), Credentials("nopass", ""))
 	if want, have := err, ErrInvalidCredentials; want != have {
 		t.Errorf("want %v, have %v", want, have)
 	}
 
-	_, err = NewClient([]string{"localhost"}, logger, ConnectTimeout(0))
+	_, err = NewClient([]string{"localhost"}, log.NewNopLogger(), ConnectTimeout(0))
 	if err == nil {
 		t.Errorf("expected connect timeout error")
 	}
 
-	_, err = NewClient([]string{"localhost"}, logger, SessionTimeout(0))
+	_, err = NewClient([]string{"localhost"}, log.NewNopLogger(), SessionTimeout(0))
 	if err == nil {
 		t.Errorf("expected connect timeout error")
 	}
@@ -96,21 +99,23 @@ func TestOptions(t *testing.T) {
 func TestCreateParentNodes(t *testing.T) {
 	payload := [][]byte{[]byte("Payload"), []byte("Test")}
 
-	c, err := NewClient([]string{"localhost:65500"}, logger)
+	c, err := NewClient([]string{"localhost:65500"}, log.NewNopLogger())
 	if err != nil {
 		t.Errorf("unexpected error: %v", err)
 	}
 	if c == nil {
-		t.Fatalf("expected new Client, got nil")
+		t.Fatal("expected new Client, got nil")
 	}
-	p, err := NewPublisher(c, "/validpath", newFactory(""), logger)
+
+	p, err := NewPublisher(c, "/validpath", newFactory(""), log.NewNopLogger())
 	if err != stdzk.ErrNoServer {
 		t.Errorf("unexpected error: %v", err)
 	}
 	if p != nil {
-		t.Errorf("expected failed new Publisher")
+		t.Error("expected failed new Publisher")
 	}
-	p, err = NewPublisher(c, "invalidpath", newFactory(""), logger)
+
+	p, err = NewPublisher(c, "invalidpath", newFactory(""), log.NewNopLogger())
 	if err != stdzk.ErrInvalidPath {
 		t.Errorf("unexpected error: %v", err)
 	}
@@ -118,31 +123,35 @@ func TestCreateParentNodes(t *testing.T) {
 	if err != stdzk.ErrNoServer {
 		t.Errorf("unexpected error: %v", err)
 	}
-	// stopping Client
+
 	c.Stop()
+
 	err = c.CreateParentNodes("/validpath")
 	if err != ErrClientClosed {
 		t.Errorf("unexpected error: %v", err)
 	}
-	p, err = NewPublisher(c, "/validpath", newFactory(""), logger)
+
+	p, err = NewPublisher(c, "/validpath", newFactory(""), log.NewNopLogger())
 	if err != ErrClientClosed {
 		t.Errorf("unexpected error: %v", err)
 	}
 	if p != nil {
-		t.Errorf("expected failed new Publisher")
+		t.Error("expected failed new Publisher")
 	}
-	c, err = NewClient([]string{"localhost:65500"}, logger, Payload(payload))
+
+	c, err = NewClient([]string{"localhost:65500"}, log.NewNopLogger(), Payload(payload))
 	if err != nil {
 		t.Errorf("unexpected error: %v", err)
 	}
 	if c == nil {
-		t.Fatalf("expected new Client, got nil")
+		t.Fatal("expected new Client, got nil")
 	}
-	p, err = NewPublisher(c, "/validpath", newFactory(""), logger)
+
+	p, err = NewPublisher(c, "/validpath", newFactory(""), log.NewNopLogger())
 	if err != stdzk.ErrNoServer {
 		t.Errorf("unexpected error: %v", err)
 	}
 	if p != nil {
-		t.Errorf("expected failed new Publisher")
+		t.Error("expected failed new Publisher")
 	}
 }
